requirements.txt
amqp==2.4.2
astroid==2.2.5
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-celery-beat==1.4.0
django-filter==2.1.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.2
flake8==3.6.0
flower==0.9.3
gunicorn==19.9.0
idna==2.8
isort==4.3.15
itypes==1.1.0
Jinja2==2.10.1
kombu==4.5.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==5.4.1
psycopg2==2.7.7
pycodestyle==2.4.0
pyflakes==2.0.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
tornado==5.1.1
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
wrapt==1.11.1
